remove unused components and code
