--- src/app/calculator/motors/motor-drive/motor-drive-graph/motor-drive-graph.component.css
@@ -1,50 +1,0 @@
-.btn {
-    padding: 6px;
-    background-color: gray;
-}
-
-.btn.btn-primary.active {
-    background-color: #145A32;
-}
-
-.btn.btn-primary.active:focus {
-    box-shadow: 0 0 0 0.2rem limegreen;
-}
-
-.legend span {
-    padding-top: 25px;
-}
-
-/* #gridToggle {
-    position: absolute;
-    z-index: 9999;
-    left: 30px
-} */
-
-.margin-top {
-    margin-top: 50px;
-}
-
-
-.expanded-chart {
-    padding: 5px;
-}
-
-.btn-container {
-    text-align: center;
-}
-
-.btn-export {
-    border-radius: 2px;
-    border: 1px solid transparent;
-    background-color: #dcdcdc;
-    opacity: 0.6;
-    font-size: 14px;
-    transition: ease-in-out 0.2s all;
-}
-
-.btn-export:hover {
-    cursor: pointer;
-    opacity: 1;
-    border-color: #212529;
-}
