--- src/app/calculator/motors/motor-drive/motor-drive-graph/motor-drive-graph.component.html
@@ -1,45 +1,0 @@
-<div class="panel-chart-container" [ngClass]="{'expanded-chart': expanded}">
-
-  <div *ngIf="expanded" class="row">
-    <div class="col-12">
-      <div class="graph-btn-container">
-        <span class="fa fa-compress" *ngIf="expanded" (click)="contractChart()" (mouseenter)="initTooltip('btnCollapseChart')" (mouseleave)="hideTooltip('btnCollapseChart')"></span>
-        <app-simple-tooltip *ngIf="expanded" [message]="'Collapse graph'" [hover]="hoverBtnCollapse" [display]="displayCollapseTooltip"
-          [adjustLeft]="20" [adjustPercent]="false"></app-simple-tooltip>
-      </div>
-    </div>
-  </div>
-
-  <div class="d-flex w-100 justify-content-center">
-    <div class="btn-group">
-      <button class="btn btn-primary" [ngClass]="{'active':selectedGraphType == 'energyUse'}" (click)="setType('energyUse')">Energy
-        Use
-      </button>
-      <button class="btn btn-primary" [ngClass]="{'active':selectedGraphType == 'energyCost'}" (click)="setType('energyCost')">Energy
-        Cost
-      </button>
-    </div>
-  </div>
-
-  <div class="d-block w-100 justify-content-center align-items-center" [ngClass]="{'h-100': !expanded, 'h-90': expanded}">
-    <div class="row m-0">
-      <div class="col-12 p-0">
-        <div #ngChart></div>
-      </div>
-    </div>
-  </div>
-
-
-  <div class="col-md-12">
-    <div class="graph-btn-container">
-      <span class="fa fa-download" (click)="downloadChart()" (mouseenter)="initTooltip('btnExportChart')" (mouseleave)="hideTooltip('btnExportChart')"></span>
-      <app-simple-tooltip [message]="'Download graph as image'" [hover]="hoverBtnExport" [display]="displayExportTooltip" [adjustLeft]="20"
-        [adjustPercent]="false"></app-simple-tooltip>
-    </div>
-    <div class="graph-btn-container">
-      <span class="fa fa-expand" *ngIf="!expanded" (click)="expandChart()" (mouseenter)="initTooltip('btnExpandChart')" (mouseleave)="hideTooltip('btnExpandChart')"></span>
-      <app-simple-tooltip *ngIf="!expanded" [message]="'Expand graph'" [hover]="hoverBtnExpand" [display]="displayExpandTooltip"
-        [adjustLeft]="59" [adjustPercent]="false"></app-simple-tooltip>
-    </div>
-  </div>
-</div>
