+++ src/app/calculator/motors/motor-drive/motor-drive.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { MotorDriveComponent } from './motor-drive.component';
 import { MotorDriveService } from './motor-drive.service';
 import { MotorDriveFormComponent } from './motor-drive-form/motor-drive-form.component';
-import { MotorDriveGraphComponent } from './motor-drive-graph/motor-drive-graph.component';
 import { MotorDriveTableComponent } from './motor-drive-table/motor-drive-table.component';
 import { MotorDriveHelpComponent } from './motor-drive-help/motor-drive-help.component';
 import { ReactiveFormsModule } from '@angular/forms';
@@ -21,7 +20,7 @@ import { SimpleTooltipModule } from '../../../shared/simple-tooltip/simple-toolt
     PercentGraphModule,
     SimpleTooltipModule
   ],
-  declarations: [MotorDriveComponent, MotorDriveFormComponent, MotorDriveGraphComponent, MotorDriveTableComponent, MotorDriveHelpComponent],
+  declarations: [MotorDriveComponent, MotorDriveFormComponent, MotorDriveTableComponent, MotorDriveHelpComponent],
   providers: [MotorDriveService],
   exports: [MotorDriveComponent]
 })
