--- src/app/phast/sankey/sankey.component.html
@@ -1,6 +1,0 @@
-<!-- <div #ngChart [attr.id]="location"></div> -->
-<div #ngChart [attr.id]="location" [ngStyle]="{'display': phast.valid.isValid? 'block': 'none'}"></div>
-<div class="w-75 ml-auto mr-auto mt-3" *ngIf="!phast.valid.isValid">
-  <app-invalid-phast [phastValid]="phast.valid"></app-invalid-phast>
-</div>
-<!-- <span #btnDownload class="sankey-chart-btn fa fa-download hide-print" (click)="downloadChart()"></span> -->
