+++ src/app/shared/phast-sankey/phast-sankey.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule, DecimalPipe } from '@angular/common';
 import { PhastSankeyComponent } from './phast-sankey.component';
 import { InvalidPhastModule } from '../../phast/invalid-phast/invalid-phast.module';
+import { SankeyService } from './sankey.service';
 
 
 
@@ -15,7 +16,8 @@ import { InvalidPhastModule } from '../../phast/invalid-phast/invalid-phast.modu
     PhastSankeyComponent
   ],
   providers: [
-    DecimalPipe
+    DecimalPipe,
+    SankeyService
   ]
 })
 export class PhastSankeyModule { }
