+++ src/index.html
@@ -133,7 +133,6 @@
 	</app-root>
 
 	<script>
-		//	var d3 = require("d3");
 		var phastAddon = require('../node_modules/amo-tools-suite/build/Release/phast.node');
 		var processHeatAddon = require('../node_modules/amo-tools-suite/build/Release/processHeat.node');
 		var psatAddon = require('../node_modules/amo-tools-suite/build/Release/psat.node');
@@ -145,12 +144,9 @@
 		var chillersAddon = require('../node_modules/amo-tools-suite/build/Release/chillers.node');
 		var sviAddon = require('../node_modules/amo-tools-suite/build/Release/svi.node');
 		var wasteWaterAddon = require('../node_modules/amo-tools-suite/build/Release/wasteWater.node');
-		var compressorAddon = require('../node_modules/amo-tools-suite/build/Release/compressorsCalc.node');
-		
+		var compressorAddon = require('../node_modules/amo-tools-suite/build/Release/compressorsCalc.node');		
 		var packageJson = require('../package.json');
 		var compressorAddon = require("../node_modules/amo-tools-suite/build/Release/compressorsCalc.node");
-	//	var autoUpdater = require('electron-updater');
-		//var json2csv = require("../node_modules/dist/json2csv.js");
 	</script>
 
 </body>
